\name{nonNAs}
\alias{nonNAs}
\title{
Return items answered by each participant}
\description{
Returns items answered by each participant}
\usage{
nonNAs(x)
}
\arguments{
  \item{x}{
SAPA data.frame}
}

\value{
\item{items}{Vector of items answered by each participant}
}
\references{
https://sapa-project.org/r}
\author{
David M. Condon \email{david@sapa-project.org}

Maintainer: Jason A. French \email{jason@sapa-project.org}}

\examples{
# items.answered <- nonNAs(sapa.table)
}

